// api/src/markdownConverter/notion/models/block-types.ts
export const enum BlockTypes {
  Text = 'text',
  Header = 'header',
  SubHeader = 'sub_header',
  SubSubHeader = 'sub_sub_header',
  NumberedList = 'numbered_list',
  BulletedList = 'bulleted_list',
  ToDo = 'to_do',
  Code = 'code',
  Callout = 'callout',
  Quote = 'quote',
  Divider = 'divider',
  Image = 'image',
  Equation = 'equation',
  Embed = 'embed',
  Video = 'video'
}